WIP adding ResourceEstimator hook to stats module
Bringing back the ol' two-interface duality, but with lazy param
handling this time, so that we aren't as sensitive to mesos module
orderings.

TODO: Update some now-failing tests, try on a real cluster..

+++ slave/tag_datadog.cpp
@@ -71,7 +71,7 @@ stats::tag_datadog::TagMode stats::tag_datadog::prepare_for_tags(
     memcpy(scratch_buffer.data(), tag_section_ptr, tag_section_size);
 
     // Move forward the data following the tag section
-    size_t after_tag_section_size = size - before_tag_section_size - tag_section_size;
+    int64_t after_tag_section_size = size - before_tag_section_size - tag_section_size;
     if (after_tag_section_size > (after_tag_section_ptr - tag_section_ptr)) {
       // memmove required: Copying between ranges which overlap within the same buffer.
       memmove(tag_section_ptr, after_tag_section_ptr, after_tag_section_size);
